requirements.txt
Django==4.1.7
python-telegram-bot==13.15
phonenumbers==8.13.6
django-phonenumber-field==7.0.2
environs==9.5.0
